// src/app/shared/components/ui/button/enums/button.enum.ts
export enum ButtonVariant {
  PRIMARY = 'primary',
  SECONDARY = 'secondary',
  SUCCESS = 'success',
  DANGER = 'danger',
  WARNING = 'warning',
  INFO = 'info',
  OUTLINE = 'outline',
  GHOST = 'ghost'
}

export enum ButtonSize {
  XS = 'xs',
  SM = 'sm',
  MD = 'md',
  LG = 'lg',
  XL = 'xl'
}

export enum ButtonType {
  BUTTON = 'button',
  SUBMIT = 'submit',
  RESET = 'reset'
}

export enum IconAlign {
  LEFT = 'left',
  RIGHT = 'right'
}

export enum IconType {
  SVG = 'svg',
  MATERIAL = 'material'
}